new price formula wip

+++ programs/tally-clob/src/instructions/add_to_balance.rs
@@ -2,8 +2,8 @@ use anchor_lang::prelude::*;
 
 use crate::{errors::TallyClobErrors, User};
 
-pub fn add_to_balance(ctx: Context<AddToBalance>, amount: f64) -> Result<()> {
-    require!(amount > 0.0, TallyClobErrors::AmountToAddTooLow);
+pub fn add_to_balance(ctx: Context<AddToBalance>, amount: u64) -> Result<()> {
+    require!(amount > 0, TallyClobErrors::AmountToAddTooLow);
 
     ctx.accounts.user.add_to_balance(amount)?;
